requirements.txt
playwright>=1.44.0
pytest-playwright>=0.5.0
pytest-xdist>=3.6.1
allure-pytest==2.13.5
allure-python-commons==2.13.5
pylint
async-generator==1.10
attrs==23.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cached-property==1.5.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==2.0.12
cryptography==44.0.1
et-xmlfile==1.1.0
Events==0.5
exceptiongroup==1.2.2
fake-useragent==1.5.1
h11==0.14.0
html5lib==1.1
idna==3.7
importlib_resources==6.4.0
iniconfig==2.0.0
isodate==0.6.1
JayDeBeApi==1.2.3
JPype1==1.5.0
lxml==5.2.2
mysql-connector-python==9.2.0
numpy==2.0.1
openpyxl==3.1.5
opensearch-dsl==2.1.0
opensearch-py==2.6.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
paramiko==3.4.0
pillow==10.4.0
pip-review==1.3.0
pixelmatch==0.3.0
platformdirs==4.2.2
pluggy==1.5.0
protobuf==5.27.2
py==1.11.0
pycparser==2.22
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
PySocks==1.7.1
pyspnego==0.11.1
pystache==0.6.5
pytest==8.3.2
pytest-check==2.3.1
pytest-repeat==0.9.3
pytest-rerunfailures==14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywinrm==0.5.0
PyYAML==6.0.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
requests_ntlm==1.3.0
six==1.16.0
smbprotocol==1.13.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sshtunnel==0.4.0
suds==1.1.2
tomli==2.0.1
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
u-msgpack-python==2.8.0
urllib3==2.5.0
urllib3-secure-extra==0.1.0
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
xmltodict==0.13.0
zeep==4.2.1
zipp==3.19.2